refactor: move TypeClassMethod to fp/meta/type_class_method.py
This is necessary to avoid circular imports:

    Kind -> TypeClassMethod
    Method -> Type -> Kind
