+++ fp/meta/__init__.py
@@ -1,9 +1,10 @@
 from .kind import Kind
-from .method import TypeClassMethod
+from .type_class_method import TypeClassMethod
 from .type import Type
 from .constructor import Constructor, Var
 from .functor import Functor, Cofunctor, Bifunctor, NFunctor, ArrowFunctor, HomFunctor
 from .monad import Monad
+from .method import Method, ClassMethod
 from .lifts import Lift
 
 __all__ = [
