+++ fp/meta/method.py
@@ -1,78 +1,68 @@
-import inspect
+from __future__ import annotations
+
+import functools
+from typing import Callable
+from .type import Type
+from .constructor import Var
+from fp.utils.exceptions import SignatureError
 
 
-class TypeClassMethod:
-    """
-    TypeClassMethod descriptor used as attribute to describe a type class.
+def _read_method_signature(annotated: Callable):
+    """Parse method signature, with first type variable substitution."""
+    annotations = annotated.__annotations__
+    tgt = annotations.pop("return")
+    src_tuple = tuple(annotations.values())
+    src0 = src_tuple[0]
+    src = src0 if len(src_tuple) == 1 else Type.Prod(*src_tuple)
+    signature = lambda T: Type.Hom(src, tgt).substitute({src0.__name__: T})
+    # Preventively evaluate signature upon method definition
+    try:
+        signature_T = signature(Var("T"))
+    except:
+        raise SignatureError(
+            f"Could not read signature from annotations: {annotations}\n\n"
+            "Make sure you annotated all variables, or use `Method.annotate` "
+            "to pass `signature : type -> Hom` explicitly."
+        )
+    return signature
 
-    Use as a decorator to declare a method for a type class `MyClass <= TypeClass`
 
-    This will register a `__dict__` lookup on type instances of `MyClass`,
-    and allows for polymorphic typing on the class methods of the type instance.
-
-    Parameters:
-    -----------
-        `signature (callable)`:
-            a function `Type -> Type` accepting a type instance as argument,
-            and returning the method's type.
-
-    Returns:
-    --------
-        `method (TypeClassMethod)`:
-            a python descriptor with `__dict__` lookup on type instances.
-            Warnings will be printed upon class instance creation
-            for any missing method implementations.
-
-    Example:
-    --------
-
-    .. code-block::
-
-        # Type class `Eq` with a single method `eq : a -> a -> Bool`
-        >>> class Eq(fp.Type):
-        ...
-        ...     @fp.TypeClassMethod
-        ...     def eq(a):
-        ...         return fp.Hom((a, a), fp.Bool)
-    """
-
-    def __set_name__(self, owner, name):
-        self._name = "_" + name
-        self.name = name
-        if hasattr(owner, "_methods_"):
-            owner._methods_ += [(name, self)]
-        else:
-            owner._methods_ = [(name, self)]
-
-    def __init__(self, signature):
-        self.signature = signature
-        self.__doc__ = signature.__doc__
-
-    def __get__(self, obj, objtype=None):
-        if obj is not None:
-            return getattr(obj, self._name)
-        else:
-            return self
-
-    def __set__(self, obj, value):
-        setattr(obj, self._name, value)
+class Method:
 
     @classmethod
-    def list(cls, objtype):
-        """
-        List method names and signatures defined on a type class.
+    def annotate(cls, signature):
+        return lambda method: cls(method, signature)
 
-        Note:
-        -----
-        TypeClassMethods appear in alphabetical order. It would be more
-        understandable if they appeared in the order they were
-        created.
-        """
-        out = []
-        methods = objtype._methods_
-        members = inspect.getmembers(objtype, lambda x: isinstance(x, cls))
-        for k, _ in methods:
-            for p, m in members:
-                if k == p:
-                    out.append((k, m))
-        return out
+    def __init__(
+        self, method: Callable, signature: Callable[type, Type.Hom] | None = None
+    ):
+        if signature is None:
+            # infer signature from annotations
+            signature = _read_method_signature(method)
+        self.signature = signature
+        self._method = method
+
+    def __set_name__(self, objtype, name):
+        self.__name__ = name
+
+    def __get__(self, obj, objtype=None) -> Type.Hom.Object:
+        if obj is None:
+            # (unbound) objtype.<method> : (objtype, *args) -> tgt
+            return self.signature(objtype)(self._method)
+        elif objtype is not None:
+            # (bound) obj.<method> : () -> tgt
+            method = self.signature(objtype)(self._method)
+            return Type.Hom.partial(method, obj)
+        else:
+            # (bound) obj.<method> : (*args) -> tgt
+            method = self.signature(type(obj))(self._method)
+            return Type.Hom.partial(method, obj)
+
+
+class ClassMethod(Method):
+
+    def __get__(self, obj, objtype=None) -> Type.Hom.Object:
+        if objtype is not None:
+            return self.signature(objtype)(self._method)
+        else:
+            return self.signature(type(obj))(self._method)
